refactor unidad 5
