+++ src/org/neosuniversity/unidad5/excepciones/Cono.java
@@ -1,4 +1,4 @@
-package org.neosuniversity.unidad5.exceptions;
+package org.neosuniversity.unidad5.excepciones;
 
 public class Cono extends Circulo implements IFiguraSolida,IFiguraUtilidades {
