+++ src/org/neosuniversity/unidad5/excepciones/FiguraException.java
@@ -1,0 +1,12 @@
+package org.neosuniversity.unidad5.excepciones;
+
+public class FiguraException extends Exception {
+
+    public FiguraException(String message) {
+        super(message);
+    }
+
+    public FiguraException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
