+++ src/org/neosuniversity/unidad5/excepciones/IFiguraUtilidades.java
@@ -1,4 +1,4 @@
-package org.neosuniversity.unidad5.exceptions;
+package org.neosuniversity.unidad5.excepciones;
 
 public interface IFiguraUtilidades {
